pyproject.toml
[tool.poetry]
name = "thesis_code"
version = "0.1.0"
description = "This project uses two models, a Masked Diffusion Language Model to generate SELFIES and a property prediction GCNN model to calculate the properties of the generated molecules."
authors = ["Daan Kaasjager <daankaasjager@hotmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
datasets = "=3.3.1"
einops = "=0.8.1"
fcd = "=1.2.2"
hydra-core = "^1.2.0"           # for hydra.utils, hydra-submitit-launcher if used as plugin
lightning = "2.5.0.post0"
scienceplots = "^2.1.1"
seaborn = "^0.13.2"
torch-geometric = "^2.6.1"
transformers = "^4.49.0"
wandb = "^0.19.11"
aiohappyeyeballs = "=2.6.1"
aiohttp = "=3.11.12"
aiosignal = "=1.4.0"
annotated-types = "=0.7.0"
antlr4-python3-runtime = "=4.9.3"
black = "=25.1.0"
cfgv = "=3.4.0"
click = "=8.1.8"
cloudpickle = "=3.1.1"
colorama = "=0.4.6"
contourpy = "=1.3.1"
cycler = "=0.12.1"
dill = "=0.3.8"
docker-pycreds = "=0.4.0"
fonttools = "=4.57.0"
frozenlist = "=1.7.0"
attrs = "^25.3.0"
build = "^1.3.0"
cachecontrol = "^0.14.3"
certifi = "^2025.7.14"
cffi = "^1.17.1"
charset-normalizer = "^3.4.2"
cleo = "^2.1.0"
crashtest = "^0.4.1"
cryptography = "^45.0.5"
distlib = "^0.4.0"
dulwich = "^0.24.1"
editables = "^0.5"
filelock = "^3.18.0"
git-filter-repo = "^2.47.0"
gitdb = "^4.0.12"
gitpython = "^3.1.45"
hatch-fancy-pypi-readme = "^25.1.0"
hatch-requirements-txt = "^0.4.1"
hatch-vcs = "^0.5.0"
hatchling = "^1.27.0"
html5lib = "^1.1"
huggingface-hub = "^0.34.3"
hydra-submitit-launcher = "^1.2.0"
identify = "^2.6.12"
idna = "^3.10"
importlib-metadata = "^8.7.0"
installer = "^0.7.0"
isort = "^6.0.1"
jaraco-classes = "^3.4.0"
jeepney = "^0.9.0"
jinja2 = "^3.1.6"
jsonschema = "^4.25.0"
keyring = "^25.6.0"
kiwisolver = "^1.4.8"
lightning-utilities = "^0.15.0"
lockfile = "^0.12.2"
markupsafe = "^3.0.2"
matplotlib = "^3.10.5"
more-itertools = "^10.7.0"
mpmath = "^1.3.0"
msgpack = "^1.1.1"
multidict = "^6.6.3"
mypy-extensions = "^1.1.0"
networkx = "^3.5"
ninja = "^1.11.1.4"
nodeenv = "^1.9.1"
numpy = "^2.3.2"
nvidia-cublas-cu12 = "^12.9.1.4"
nvidia-cuda-cupti-cu12 = "^12.9.79"
nvidia-cuda-nvrtc-cu12 = "^12.9.86"
nvidia-cuda-runtime-cu12 = "^12.9.79"
nvidia-cudnn-cu12 = "^9.11.0.98"
nvidia-cufft-cu12 = "^11.4.1.4"
nvidia-cufile-cu12 = "^1.14.1.1"
nvidia-curand-cu12 = "^10.3.10.19"
nvidia-cusolver-cu12 = "^11.7.5.82"
nvidia-cusparse-cu12 = "^12.5.10.65"
nvidia-cusparselt-cu12 = "^0.7.1"
nvidia-nccl-cu12 = "^2.27.6"
nvidia-nvjitlink-cu12 = "^12.9.86"
nvidia-nvtx-cu12 = "^12.9.79"
omegaconf = "^2.3.0"
pandas = "^2.3.1"
pathlib = "^1.0.1"
pathspec = "^0.12.1"
pexpect = "^4.9.0"
pillow = "^11.3.0"
pip-chill = "^1.0.3"
pipdeptree = "^2.28.0"
pkginfo = "^1.12.1.2"
platformdirs = "^4.3.8"
pluggy = "^1.6.0"
poetry-core = "^2.1.3"
pre-commit = "^4.2.0"
propcache = "^0.3.2"
protobuf = "^6.31.1"
psutil = "^7.0.0"
ptyprocess = "^0.7.0"
pyarrow = "^21.0.0"
pycparser = "^2.22"
pydantic = "^2.11.7"
pydantic-core = "^2.37.2"
pyparsing = "^3.2.3"
pyproject-hooks = "^1.2.0"
pyrsistent = "^0.20.0"
python-dateutil = "^2.9.0.post0"
pytorch-lightning = "^2.5.2"
pytz = "^2025.2"
pyyaml = "^6.0.2"
rapidfuzz = "^3.13.0"
rdkit = "^2025.3.5"
regex = "^2025.7.34"
requests = "^2.32.4"
requests-toolbelt = "^1.0.0"
safetensors = "^0.5.3"
scipy = "^1.16.1"
secretstorage = "^3.3.3"
selfies = "^2.2.0"
sentry-sdk = "^2.34.1"
setproctitle = "^1.3.6"
shellingham = "^1.5.4"
six = "^1.17.0"
smmap = "^5.0.2"
submitit = "^1.5.3"
sympy = "^1.14.0"
tokenizers = "^0.21.4"
tomlkit = "^0.13.3"
torchmetrics = "^1.8.0"
tqdm = "^4.67.1"
trove-classifiers = "^2025.5.9.12"
typing-extensions = "^4.14.1"
tzdata = "^2025.2"
urllib3 = "^2.5.0"
virtualenv = "^20.32.0"
webencodings = "^0.5.1"
xxhash = "^3.5.0"
yarl = "^1.20.1"
zipp = "^3.23.0"

[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"
black = "^25.1.0"
ruff = "^0.11.7"
pip-chill = "^1.0.3"
pip-tools = "^7.4.1"
pipdeptree = "^2.26.1"

[[tool.poetry.packages]]
include = "common"
from    = "src"

[[tool.poetry.packages]]
include = "mdlm"
from    = "src"

[[tool.poetry.packages]]
include = "property_prediction"
from    = "src"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
